Concedendo somente aos 'admins' o poder e inativação/ativação de registros

+++ app/clientes/dClientes.php
@@ -2,6 +2,10 @@
 session_start();
 include '../config/conexao.php';
 
+if($_SESSION['cargo'] !== 'admin'){
+    die("Acesso negado! Apenas administradores podem executar esta ação.");
+}
+
 $id = $_GET['id'];
 
 $sql_delete_cliente = "DELETE FROM clientes WHERE id = $id";
